pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-api-python-client>=2.164.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "flask-wtf>=1.2.2",
    "sqlalchemy>=2.0.39",
]
